自定义房间，并且真人和Ai都可以进房间

+++ Assets/Scripts/Event/NgoLoadSceneCompleteEvent.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 
 public class NgoLoadSceneCompleteEvent : EventBase
 {
-    public ulong ClientId;
+    public int ClientId;
     public string SceneName;
 }
